ai/requirements.txt
flask
flask-cors
groq
google-genai
firebase-admin
python-dotenv